refactor: split native command orchestration
